requirements.txt
absl-py==2.1.0
accelerate==1.5.2
addict==2.4.0
asttokens==3.0.0
av==14.2.0
bitsandbytes==0.45.3
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.56.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
grpcio==1.71.0
huggingface-hub==0.29.3
icecream==2.1.4
identify==2.6.9
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
pre_commit==4.2.0
protobuf==6.30.1
psutil==7.0.0
Pygments==2.19.1
pyiqa==0.1.13
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tifffile==2025.3.13
timm==1.0.15
tokenizers==0.15.2
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.37.2
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
wcwidth==0.2.13
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
